Update optimization.rst
It is really really well explained.

A minor question, apart from the changes. Does that  'n jobs': int(multiprocessing.cpu_count()/2) works as a universal principle? I think it works nicely for all standard intel processors (checked it on i5 and i7 for varoius versions), but I am not sure it works the same way for other types e.g. Xeon. I recently launched a job (using joblib package though) on the full number of cpu_count result and it scaled quite well. However, I have limited knowledge on the reason that scaled that nicely.
I suggest to avoid giving a specific guideline on the selection through cpu_count. It could be mentioned as a suggestion to obtain more information on cpu/thread counts, while the final decision shall be made by the user.